added footer, moved github profile to footer & added linkedin/stackoverflow
